requirements.txt
asgiref==3.8.1
coverage==7.6.4
Django==5.1.2
django-allauth==65.1.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
